Корректировка архетиктуры
