+++ CMakelists.txt
@@ -3,19 +3,20 @@ cmake_minimum_required (VERSION 3.8)
 project (project)
 
 add_executable (sea_battle
-	src/main.cpp
-    src/Ship.cpp
-    src/Player.cpp
- )
+	main.cpp
+    #src/Ship.cpp
+    #src/Player.cpp
+    #src/Game.cpp
+    )
 
 target_include_directories(
     sea_battle
-   PRIVATE ${CMAKE_CURRENT_BINARY_DIR}
+   PRIVATE ${CMAKE_CURRENT_BINARY_DIR}/Game
    )
 
 target_include_directories(
     sea_battle 
-    PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/headers
+    PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Game/headers
     )
 
 set_target_properties (sea_battle PROPERTIES
