--- headers/Game.h
@@ -1,124 +1,0 @@
-#pragma once
-#include <iostream>
-
-#include <string>
-#include <vector>
-#include <memory>
-#include <stdlib.h>
-#include <ctime>
-#include "Player.h"
-
-//static enum direction{
-//	UP=1,
-//	RIGHT,
-//	DOWN,
-//	LEFT
-//} dir;
-
-class Game {
-private:
-	
-	std::unique_ptr<Player> plr1;
-	std::unique_ptr<Player> plr2;
-	int pol;
-	bool gm_over=false;	//Проверка конца игры
-	bool input_mode=false;	//true-ручной, false- автоматический
-	
-
-public:
-	//Конструктор (задает способ расстановки и размер поля)
-	Game(bool m, int p=10) :plr1(std::make_unique<Player>(10)),
-		plr2(std::make_unique<Player>(10)), pol(p), input_mode(m){
-		if (input_mode) {
-			manSetShip();
-			plr1 = autoSetShip(std::move(plr1));
-		}
-		else {
-			plr1 = autoSetShip(std::move(plr1));
-			plr2 = autoSetShip(std::move(plr2));
-		}
-	};
-	
-	//Ручная расстановка кораблей
-	void manSetShip() {
-		int x = 0;
-		int y = 0;
-		int d = 0;
-		std::cout << "Ручной ввод расположения кораблей:"
-			<< std::endl;
-		for (int i = 4; i > 0; --i) {
-			for (int j = 4; j >= i; --j) {
-				std::cout << "Введите координаты и направление " << j << "-х палубного корабля"
-					<< std::endl;
-				std::cout << "Направление (1-вверх,2-вправо,3-вниз,4-влево): " << std::endl;	
-				for (;;) {
-					std::cout << "x,y,dir: ";
-					std::cin >> x >> y >> d;
-					if (testCords(x, y, d, i)) {
-						std::cout << "Выход за пределы поля! Попробуйте снова" << std::endl;
-						continue;
-					}
-					if (!plr1->setShip(x, y, d, i)) {
-						std::cout << "Есть пересечение с лругим кораблем! Попробуйте снова."
-							<< std::endl;
-						continue;
-					}
-					else
-						break;
-				}
-			}
-		}
-	}
-	// Автоматическая расстановка кораблей
-	std::unique_ptr<Player>&& autoSetShip(std::unique_ptr<Player>&& pl) {
-		std::srand(static_cast <unsigned int>(std::time(nullptr)));
-		int x = 0;
-		int y = 0;
-		int d = 0;
-		for (int i = 4; i > 0; --i) {
-			for (int j = 4; j >= i; --j) {
-				for (;;) {
-					x = std::rand() % 10 + 1;
-					y = std::rand() % 10 + 1;
-					d = std::rand() % 4 + 1;
-					if (testCords(x, y, d, i))
-						continue;
-					if (!pl->setShip(x, y, d, i))
-						continue;
-					else
-						break;
-				}
-			}
-		}
-		return std::move(pl);
-	}
-	//Проверка выхода за границу поля
-	bool testCords(int& _x, int& _y, int& _dir, int& _deck) {
-		if (_x <= 0 || _y <= 0 || _x > pol || _y > pol)
-			return true;
-		if (_dir == 1)
-			return ((_x - _deck) <= 0);
-		else if (_dir == 2)
-			return ((_y + _deck) > pol);
-		else if (_dir == 3)
-			return ((_x - _deck) > pol);
-		else 
-			return ((_y - _deck) <= 0);
-	}
-	//Проверка наличия кораблей
-	bool isOver() {
-		if (!plr1->ShipCount()){
-			std::cout << "Игрок 2 выиграл!" << std::endl;
-			return true;
-		}
-		if (!plr2->ShipCount()) {
-			std::cout << "Игрок 1 выиграл!" << std::endl;
-			return true;
-		}
-		return false;
-	}
-
-
-
-
-};
